hw5: add abstract repo tests and updaters chain
