+++ bot/src/main/java/edu/java/bot/configuration/LinkValidatorConfig.java
@@ -4,11 +4,9 @@ import edu.java.bot.service.link.AbstractLinkValidator;
 import edu.java.bot.service.link.LinkUtils;
 import java.util.Set;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan(basePackages = "edu.java.bot.service.link")
 public class LinkValidatorConfig {
 
     private final Set<AbstractLinkValidator> linkValidatorSet;
