+++ scrapper/src/main/java/edu/java/scrapper/api/service/updater/AbstractUpdatesFetcher.java
@@ -9,6 +9,7 @@ import edu.java.scrapper.api.service.exception.EntityNotFoundException;
 import edu.java.scrapper.api.service.exception.NotSupportedLinkException;
 import java.net.URI;
 import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,16 +18,38 @@ public abstract class AbstractUpdatesFetcher {
     private final SubscriptionRepository subscriptionRepository;
 
     private final LinkRepository linkRepository;
+    private AbstractUpdatesFetcher next;
 
     protected AbstractUpdatesFetcher(SubscriptionRepository subscriptionRepository, LinkRepository linkRepository) {
         this.subscriptionRepository = subscriptionRepository;
         this.linkRepository = linkRepository;
     }
 
+    public void setNext(AbstractUpdatesFetcher next) {
+        if (this.equals(next)) {
+            throw new IllegalArgumentException("В цепочке не могут состоять одинаковые объекты");
+        }
+        this.next = next;
+    }
+
+    public Optional<LinkUpdateRequest> chainedUpdatesFetching(LinkDto linkDto)
+        throws NotSupportedLinkException, EntityNotFoundException {
+        try {
+            return fetchUpdatesFromLink(linkDto);
+        } catch (NotSupportedLinkException e) {
+            if (next != null) {
+                return next.fetchUpdatesFromLink(linkDto);
+            } else {
+                throw e;
+            }
+        }
+    }
+
     public abstract Optional<LinkUpdateRequest> fetchUpdatesFromLink(LinkDto linkDto)
         throws NotSupportedLinkException, EntityNotFoundException;
 
     protected String[] getUrlPathParts(URI url, String siteBaseUrl) throws NotSupportedLinkException {
+        FetchersChainUtils.throwIfProtocolAbsent(url.toString());
         if (!siteBaseUrl.equals(url.getHost())) {
             throw new NotSupportedLinkException("Сервис %s не поддерживается".formatted(url.getHost()));
         }
@@ -44,13 +67,18 @@ public abstract class AbstractUpdatesFetcher {
         return substring.split(URL_DELIMITER);
     }
 
-    protected Optional<LinkUpdateRequest> defineShouldMakeUpdateRequest(
+    protected Optional<LinkUpdateRequest> makeLinkUpdate(
         LinkDto linkDto,
         OffsetDateTime fetchedUpdateDate,
         String description
     ) {
-        OffsetDateTime checkedTime = OffsetDateTime.now();
-        if (linkDto.lastUpdatedAt().isAfter(fetchedUpdateDate)) {
+        OffsetDateTime checkedTime = OffsetDateTime.now(ZoneOffset.UTC);
+        OffsetDateTime lastUpdatedAt = linkDto.lastUpdatedAt();
+
+        boolean dontNeedUpdate = lastUpdatedAt != null
+            && (fetchedUpdateDate.isEqual(lastUpdatedAt) || fetchedUpdateDate.isBefore(lastUpdatedAt));
+
+        if (dontNeedUpdate) {
             return Optional.empty();
         }
