+++ scrapper/src/main/java/edu/java/scrapper/api/service/updater/SimpleLinkUpdaterService.java
@@ -7,14 +7,15 @@ import edu.java.scrapper.api.repository.dto.ChatDto;
 import edu.java.scrapper.api.repository.dto.LinkDto;
 import edu.java.scrapper.api.service.exception.EntityNotFoundException;
 import edu.java.scrapper.api.service.exception.NotSupportedLinkException;
-import edu.java.scrapper.api.service.updater.github.GithubUpdatesFetcher;
 import edu.java.scrapper.client.bot.BotClient;
 import edu.java.scrapper.configuration.ApplicationConfig;
 import java.time.Duration;
 import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,26 +25,27 @@ public class SimpleLinkUpdaterService implements LinkUpdaterService {
     private final SubscriptionRepository subscriptionRepository;
     private final ApplicationConfig config;
     private final BotClient botClient;
-    private final GithubUpdatesFetcher githubUpdatesFetcher;
+    private final AbstractUpdatesFetcher headUpdatesFetcher;
 
     public SimpleLinkUpdaterService(
         LinkRepository linkRepository,
-        SubscriptionRepository subscriptionRepository, ApplicationConfig config, BotClient botClient,
-        GithubUpdatesFetcher githubUpdatesFetcher
+        SubscriptionRepository subscriptionRepository,
+        ApplicationConfig config,
+        BotClient botClient,
+        @Qualifier("headUpdatesFetcher") AbstractUpdatesFetcher headUpdatesFetcher
     ) {
         this.linkRepository = linkRepository;
         this.subscriptionRepository = subscriptionRepository;
         this.config = config;
         this.botClient = botClient;
-        this.githubUpdatesFetcher = githubUpdatesFetcher;
+        this.headUpdatesFetcher = headUpdatesFetcher;
     }
 
     @Override
     @Transactional
     public int update() {
         Duration delay = config.scheduler().forceCheckDelay();
-
-        OffsetDateTime limitTime = OffsetDateTime.now().minus(delay);
+        OffsetDateTime limitTime = OffsetDateTime.now(ZoneOffset.UTC).minus(delay);
 
         long linksLimit = config.scheduler().scanLinksLimit();
         Collection<LinkDto> allBeforeLastSchedulerCheck =
@@ -53,13 +55,9 @@ public class SimpleLinkUpdaterService implements LinkUpdaterService {
         for (LinkDto linkDto : allBeforeLastSchedulerCheck) {
 
             Optional<LinkUpdateRequest> requestOptional;
+
             try {
-                try {
-                    requestOptional = githubUpdatesFetcher.fetchUpdatesFromLink(linkDto);
-                } catch (NotSupportedLinkException e) {
-                    // TODO stackoverflow
-                    requestOptional = githubUpdatesFetcher.fetchUpdatesFromLink(linkDto);
-                }
+                requestOptional = headUpdatesFetcher.chainedUpdatesFetching(linkDto);
             } catch (NotSupportedLinkException | EntityNotFoundException e) {
                 processInvalidLink(linkDto, e.getMessage());
                 continue;
@@ -69,18 +67,6 @@ public class SimpleLinkUpdaterService implements LinkUpdaterService {
                 countUpdates++;
                 botClient.updates(requestOptional.get());
             }
-
-            /*OffsetDateTime fetchedUpdateDate = null;
-            if (linkDto.lastUpdatedAt().isBefore(fetchedUpdateDate)) {
-                List<Long> chatIds = subscriptionRepository.findChatsByLinkId(linkDto.id()).stream()
-                    .map(ChatDto::id)
-                    .toList();
-
-                LinkUpdateRequest linkUpdateRequest =
-                    new LinkUpdateRequest(linkDto.id(), linkDto.url(), "Информация обновилась", chatIds);
-                botClient.updates(linkUpdateRequest);
-            }*/
-
         }
         return countUpdates;
     }
@@ -93,7 +79,7 @@ public class SimpleLinkUpdaterService implements LinkUpdaterService {
             new LinkUpdateRequest(
                 linkDto.id(),
                 linkDto.url(),
-                "Ссылка будет удалена из отслеживаемых. Причина:\n" + errorMessage,
+                "Ссылка %s будет удалена из отслеживаемых. Причина:\n%s".formatted(linkDto.url(), errorMessage),
                 chatIds
             );
         botClient.updates(notUpdateRequest);
