+++ scrapper/src/main/java/edu/java/scrapper/sheduler/LinkUpdaterScheduler.java
@@ -1,5 +1,6 @@
 package edu.java.scrapper.sheduler;
 
+import edu.java.scrapper.api.service.updater.LinkUpdaterService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -10,8 +11,15 @@ import org.springframework.stereotype.Component;
 @ConditionalOnProperty(value = "app.scheduler.enable")
 public class LinkUpdaterScheduler {
 
+    private final LinkUpdaterService linkUpdaterService;
+
+    public LinkUpdaterScheduler(LinkUpdaterService linkUpdaterService) {
+        this.linkUpdaterService = linkUpdaterService;
+    }
+
     @Scheduled(fixedDelayString = "${app.scheduler.interval}")
     public void update() {
-        log.info("LinkUpdaterScheduler выполнил метод update");
+        int countUpdates = linkUpdaterService.update();
+        log.info("LinkUpdaterScheduler обновил {} ссылок", countUpdates);
     }
 }
